Stop managing focus unless item is menu item
The current iteration of the code is re-focusing items
based on the currently active element being either a menu
item or the dropdown trigger itself.

This causes a Tab-trap, where a user who tabs backwards
(using shift-tab) ending up on the dropdown element
automatically triggers a focusing of the first dropdown
item.

This change should fix this issue, while still allowing for
arrow key navigation. It does also prevent the first item
from being focused when the dropdown is opened,
but that is (IMHO) desired behavior, and in line with
how official Bootstrap handles this.