Implement dismissal of dropdown when focus outside
If the user focuses an item outside of the dropdown, the dropdown should close (this is how original Bootstrap dropdown works).

The implementation uses a focus listener delegated to the document. Delegating the focus event requires use of the useCapture flag, since focus/blur do not bubble.

If the focused event is the menu, the button triggering the menu, or originates inside the menu, the dropdown is not dismissed.