Fix over-eager focusing of trigger button
The button that triggers the dropdown is focused on close,
no matter how the dropdown was closed.

The original behavior causes focusing of the trigger for example even when the user clicks another focusable item on the page, potentially competing for focus management.

The current solution checks if the menu contains the activeElement of the document, and only then focuses the trigger, which seems to me to be the desired behavior.