Refactor delegation to allow for passing options
Event delegation mostly works with options|useCapture set to false.
Delegating focus requires useCapture flag to be truthy,
so I've added the option to pass in options, so to speak.

(it used to be a boolean flag, but newer DOM standards allow an options object, e.g. {once: true, useCapture: true} etc.)